requirements.txt
Django==1.5.4
MySQL-python==1.2.5
Pillow==2.3.0
South==0.8.4
agon-ratings==0.2
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.3.1
billiard==3.3.0.17
bitly-api==0.3
boto==2.27.0
celery==3.1.11
dj-database-url==0.2.2
dj-static==0.0.5
django-allauth==0.13.0
django-appconf==0.6
django-audiofield==0.6.2
django-autocomplete-light==1.4.7
django-bootstrap-form==2.0.6
django-braces==1.4.0
django-brake==1.3.1
django-celery==3.1.10
django-ckeditor==4.0.2
django-classy-tags==0.5
django-contact-form==1.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.0.1
django-extensions==1.4.6
django-filebrowser==3.5.2
django-filepicker==0.1.4
django-filer==0.9.4
django-filter==0.7
django-forms-bootstrap==3.0.0
django-grappelli==2.4.6
django-guardian==1.2.0
django-haystack==2.1.0
django-jsonfield==0.9.10
django-localflavor==1.0
django-locality==0.1.1
django-mptt==0.6.0
django-parsley==0.0.2a0
django-polymorphic==0.4.2
django-pylibmc==0.5.0
django-pylibmc-sasl==0.2.4
django-sekizai==0.7
django-storages==1.1.8
django-stripe-payments==2.0b34
django-suit==0.2.2
django-toolbelt==0.0.1
djangorestframework==2.3.12
djrill==0.9.0
docopt==0.4.0
drf-ujson==1.1
dropbox==1.5.1
easy-thumbnails==1.2
elasticsearch==0.4.3
flower==0.6.0
fudge==1.0.3
geturl==0.2.1
gevent==1.0.1
greenlet==0.4.4
grequests==0.2.0
gunicorn==18.0
hiredis==0.1.1
html5lib==0.99
httplib2==0.9
ipython==2.0.0
kombu==3.0.16
lxml==3.2.3
mandrill==1.0.53
mechanize==0.2.5
micawber==0.3.0
nodeenv==0.7.3
nose==1.3.4
oauth2==1.5.211
oauthlib==0.6.0
phileo==1.1
ply==3.4
psycopg2==2.5.2
pyelasticsearch==0.6.1
pylibmc==1.2.3
pystache==0.5.3
python-dateutil==2.1
python-memcached==1.53
python-openid==2.2.5
python-social-auth==0.1.23
pytz==2014.2
recaptcha-client==1.0.6
redis==2.9.1
requests==1.2.3
requests-oauthlib==0.3.2
simplejson==3.3.1
slimit==0.8.1
soundcloud==0.4.1
sqlparse==0.1.11
static==1.0.2
stripe==1.15.1
tornado==3.1.1
twilio==3.6.6
twython==3.0.0
uWSGI==2.0.5.1
ujson==1.33
urllib3==1.7.1
wsgiref==0.1.2
python-twitter
django-sslify>=0.2.0
rsa
retrying
six==1.7.0
django-embed-video
